Optimizing performance and readability of the Grid System code
In this commit I have optimized the performance and readability of the Grid System code. In the GameManager I changed the code so Update Placement no longer loops through all of the GridCells in the scene, and instead I am using a line trace with a Custom made Trace Channle that I called GridCell.