Fixed bug with the highlight material not showing when going into Deletion Mode while already being hovered over the machine
What I have done is, I created an Interface, 'Toggle Mode', and sent a message from the GameManager, when the game goes or exits deletion mode, to the Clickable Component. Then I created an Object Channel 'PlaceableActor' that I set the BaseBuildingAsset to.
Then the code in the clickable component goes as follows: "Get the interface for when the game goes into the deletion mode alongside the 'DeletionMode' bool; fire a ray cast from the mouse position towards the floor, make the raycast only check for actors with ObjectChannel 'PlaceableActor', then get the HitActor's static meshes and skeletal meshes, add them to the 'StaticMeshToSetHighlightTo' array, then loop through the array and set the CustomDepth to true or false based on the 'DeletionMode' boolean we got through the interface"